// src/app/model/catdoccandidato.model.ts
export interface Catdoccandidato {
  cndIdcandidato: number;
  cndNombre: string;
  cndApellidoPat: string;
  cndApellidoMat: string;
  cndCalle: string;
  cndNumext: string;
  cndNumint: string;
  cndCodpostal: string;
  cndColonia: string;
  cndMunicipio: string;
  cndEstado: string;
  cndPais: string;
  cndEdad: string;
  cndEstcivil: string;
  cndEmail: string;
  cndHijos: string;
  cndDispoviajar: string;
  cndVisa: string;
  cndPasaporte: string;
  cndTrabaja: string;
  cndDisponibildiad: string;
  cndDispotrabajo: string;
  cndGenero: string;
}